app/requirements.txt
aiohttp==3.7.3
amqp==5.0.5
appdirs==1.4.4
appnope==0.1.2
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.4
decorator==4.4.2
demjson==2.2.4
Django==3.1.6
django-advanced-filters==1.2.0
django-braces @ https://github.com/galuszkak/django-braces/archive/5c6604123eaa73f53875c5dce5843a6598b7b52c.zip
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-extensions==3.1.1
docopt==0.6.2
ecdsa==0.16.1
idna==2.10
ipdb==0.13.4
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
lxml==4.6.2
ms-cv==0.1.1
multidict==5.1.0
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
PyCap==1.1.2
pycparser==2.20
pycrypto==2.6.1
pydantic==1.7.3
Pygments==2.8.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pysftp==0.2.9
pyTelegramBotAPI==3.7.6
pytz==2021.1
redcap-cli==0.1.0
redi==0.15.5
redis==3.5.3
requests==2.25.1
semantic-version==2.8.5
setuptools-scm==5.0.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
urwid==2.1.2
vine==5.0.0
wcwidth==0.2.5
xbox-webapi-ex==1.1.9
yarl==1.6.3
